backend/requirements.txt
fastapi
uvicorn[standard]
flask
flask-cors
numpy
pandas
scikit-learn
tensorflow
python-multipart
opencv-python
pillow
fpdf2
openai
python-dotenv
